lambda-test-project/requirements.txt
selenium==4.19.0
pytest
python-dotenv
